Interviewer_model/requirements.txt
# AI Interview Coach - Requirements
# Python 3.9+

# ===== Core Deep Learning =====
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0  # 4bit量化支持

# ===== Speech Processing =====
openai-whisper>=20231117
librosa>=0.10.0
soundfile>=0.12.0
pyaudio>=0.2.13  # 实时音频输入
wave>=0.0.2

# ===== NLP & Embeddings =====
sentence-transformers>=2.2.2
numpy>=1.24.0
scikit-learn>=1.3.0

# ===== Document Processing =====
pdfplumber>=0.10.0
python-docx>=1.0.0
pypdf2>=3.0.0

# ===== UI & Visualization =====
streamlit>=1.28.0
streamlit-webrtc>=0.47.0  # 网页实时音视频
plotly>=5.17.0
pillow>=10.0.0

# ===== Data & Utils =====
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0

# ===== Optional: Training =====
datasets>=2.14.0
peft>=0.6.0  # LoRA微调
tensorboard>=2.15.0

# ===== Development =====
pytest>=7.4.0
black>=23.10.0
flake8>=6.1.0

